fix(ci): python-coverage-shards — extract pre-push validation test helpers to stay under file line limit
The PRRT_kwDOSJAM6s6HjHiR review-comment fix pushed
tests/unit/runtime/test_pr_monitor_pre_push_validation.py to 1525 lines,
breaching the 1500-line first-party file guardrail enforced by
test_core_decomposition_maintainability.py. Move the shared validation
builders, fake runner, and seeding fixtures into a private
_pre_push_validation_helpers module and import them, dropping the test
module to 1289 lines without weakening the guardrail.
